requirements.txt
pymongo[srv]
python-telegram-bot
aiohttp
flask
python-dotenv
gunicorn